Fix: Bill denomination type changed from double to int

+++ lib/Model/Bill.dart
@@ -1,5 +1,5 @@
 class Bill {
-  double denomination;
+  int denomination;
   int amount;
 
   Bill(this.denomination, this.amount);
